+++ src/lib/ReactDataTable/ReactDataTableProps.ts
@@ -113,4 +113,9 @@ export interface ReactDataTableProps<TData, TFilter extends FilterModel> {
    * @default true
    */
   showClearSearchButton?: boolean;
+
+  /**
+   * A component to render as a sub-row for a specific row
+   */
+  subRowComponent?: (row: Row<TData>) => React.ReactNode;
 }